refactor(sys): reduce build dependencies
This shaves 2 seconds off `cargo build --example model-info` (a basic binary with only one dependency on `ort`).

Note that `pkg-config` currently requires a feature flag to enable. In the future let's just use the `pkg-config` binary directly.

On Windows, Ryzen 5800X:
```
         base: 75 crates in 12.94s
/ hmac-sha256: 64 crates in 12.89s
/     zlib-rs: 64 crates in 11.47s
-  pkg-config: 63 crates in 11.72s
-         tar: 61 crates in 10.34s
```

These gains are probably less substantive on Linux/macOS, since a good chunk of the prior build time was due to the inclusion of both `windows-sys` and `winapi` in the dependency tree.